requirements.txt
flask==3.0.0
openai==1.12.0
gspread==6.0.2
oauth2client==4.1.3
pandas==2.2.0
python-dotenv==1.0.1
